Move common dummies to common file.
