+++ test/hmi/test_dummies.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <board_types.hpp>
+#include <functional>
+#include <map>
+
+extern std::map<board::PinType, std::function<void(void)>> isr_collection;
